xmly/requirements.txt
pyppeteer
opencv-python
Pillow
requests